move tmtv
